add play button component
